Return zero for all neutrino types which are unsupported by model
The old behavior was to throw on particle type which was not included in the
model. This was causing problems with tau neutrinos where the flux is expected
to be either zero or significantly lower than the other flavors.
This changes it so that the six neturino types will never throw.
The value of zero is used for any particle type wihch is not in the model.

Fixes #15