Switch Python build system to use mesonpy instead of setuptools (#24)
This PR solves a couple of problems at once. I got an email from a user complaining that the setuptools build system couldn't handle multiple photospline and boost being in different directories. Meson's pkg-config dependency detection is far superior to anything you can ad hoc program in setuptools. This allowed some cleanup that I had wanted to do regarding the meson build file. Also, one of the reasons that I gave up on making wheels for nuflux was that setuptools couldn't figure out how to include header files while mesonpy does this automatically. The files changed looks big but that is just because I also moved the python source directory to /src/.

The only change from the user perspective is that when compiling you now use exactly the same env vars for both meson and python. Documentation has been updated to reflect this. cvmfs target py3-v4.0.1 was dropped because the last version of meson that worked with python 3.6 doesn't support the pure option for python. And py3-v4.2.1 was dropped because mesonpy doesn't play well with the python misconfiguration present in that version of cvmfs.