unset PYTHONPATH in cvmfs tests (#30)
* unset PYTHONPATH in cvmfs tests

PYTHONPATH appears to be not only unnescessary but activly messing up the tests
because the packages in system site-packages are old and incompatible with the
latest version of pytest. Having system site-packages in PYTHONPATH causes
them to be loaded first overriding what it is in .local/

* remove v4.2.1 from pseudo-cobalt gh action

There is some binary incompatability issue with running the latest wheels from
pypi on top of cvmfs python on top of the centos docker image that only applies
to v4.2.x. I don't want to spend any more time trying to figure out. The
testing regime is already pretty robust. So I am just gonna drop v4.2.1 and
move on.